[ET][Portable][Build Size] Move to dtype_utils
Differential Revision: [D63994875](https://our.internmc.facebook.com/intern/diff/D63994875/)

[ghstack-poisoned]
